backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.21
pymysql==1.1.0
PyJWT==2.8.0
numpy==1.26.3
requests==2.31.0
